+++ packages/gnode/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @happyvertical/smrt-gnode
 
+## 0.40.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.17
+
+## 0.40.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.16
+
 ## 0.40.15
 
 ### Patch Changes