+++ packages/manufacturing/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @happyvertical/smrt-manufacturing
 
+## 0.40.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.17
+  - @happyvertical/smrt-inventory@0.40.17
+  - @happyvertical/smrt-tenancy@0.40.17
+
+## 0.40.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.16
+  - @happyvertical/smrt-inventory@0.40.16
+  - @happyvertical/smrt-tenancy@0.40.16
+
 ## 0.40.15
 
 ### Patch Changes