+++ packages/reports/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @happyvertical/smrt-reports
 
+## 0.40.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.17
+  - @happyvertical/smrt-jobs@0.40.17
+  - @happyvertical/smrt-tenancy@0.40.17
+
+## 0.40.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.16
+  - @happyvertical/smrt-jobs@0.40.16
+  - @happyvertical/smrt-tenancy@0.40.16
+
 ## 0.40.15
 
 ### Patch Changes