+++ packages/smrt-mobile-contract/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @happyvertical/smrt-mobile-contract
 
+## 0.40.17
+
+## 0.40.16
+
 ## 0.40.15
 
 ## 0.40.14