+++ packages/users/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @happyvertical/smrt-users
 
+## 0.40.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.17
+  - @happyvertical/smrt-profiles@0.40.17
+  - @happyvertical/smrt-mobile-contract@0.40.17
+  - @happyvertical/smrt-tenancy@0.40.17
+  - @happyvertical/smrt-config@0.40.17
+  - @happyvertical/smrt-types@0.40.17
+  - @happyvertical/smrt-ui@0.40.17
+
+## 0.40.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.16
+  - @happyvertical/smrt-profiles@0.40.16
+  - @happyvertical/smrt-mobile-contract@0.40.16
+  - @happyvertical/smrt-tenancy@0.40.16
+  - @happyvertical/smrt-config@0.40.16
+  - @happyvertical/smrt-types@0.40.16
+  - @happyvertical/smrt-ui@0.40.16
+
 ## 0.40.15
 
 ### Patch Changes