+++ packages/video/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @happyvertical/smrt-video
 
+## 0.40.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.17
+  - @happyvertical/smrt-assets@0.40.17
+  - @happyvertical/smrt-content@0.40.17
+  - @happyvertical/smrt-profiles@0.40.17
+  - @happyvertical/smrt-tenancy@0.40.17
+  - @happyvertical/smrt-voice@0.40.17
+  - @happyvertical/smrt-config@0.40.17
+
+## 0.40.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.40.16
+  - @happyvertical/smrt-assets@0.40.16
+  - @happyvertical/smrt-content@0.40.16
+  - @happyvertical/smrt-profiles@0.40.16
+  - @happyvertical/smrt-tenancy@0.40.16
+  - @happyvertical/smrt-voice@0.40.16
+  - @happyvertical/smrt-config@0.40.16
+
 ## 0.40.15
 
 ### Patch Changes